% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{clean_parameters}
\alias{clean_parameters}
\title{Clean parameters list}
\usage{
clean_parameters(params)
}
\arguments{
\item{params}{named list; parameter set}
}
\value{
parameter set
}
\description{
The parameters that are read in from \code{params.yaml} needs a few tweaks before
they can get passed into the main simulation function
}
