% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_params.R
\name{date_to_day}
\alias{date_to_day}
\title{Date to integer day index}
\usage{
date_to_day(dt, start_date)
}
\arguments{
\item{dt}{date}

\item{start_date}{day 1}
}
\value{
integer index
}
\description{
Date to integer day index
}
