% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_params.R
\name{day_to_date}
\alias{day_to_date}
\title{Integer day index to date}
\usage{
day_to_date(day, start_date)
}
\arguments{
\item{day}{integer day}

\item{start_date}{date of day 1}
}
\value{
date
}
\description{
Integer day index to date
}
