% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{do_importations}
\alias{do_importations}
\title{Do importations}
\usage{
do_importations(x, imports)
}
\arguments{
\item{x}{simulation compartment state}

\item{imports}{vector of importations, with length equal to number of
populations}
}
\value{
state vector
}
\description{
Do importations
}
