% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{do_vaccinations}
\alias{do_vaccinations}
\title{Do vaccinations}
\usage{
do_vaccinations(x, doses, ve, ve_infant, prop_vax_to_se)
}
\arguments{
\item{x}{simulation compartment state}

\item{doses}{integer number of dose counts}

\item{ve}{non-infant vaccine effectiveness}

\item{ve_infant}{infant vaccine effectiveness}

\item{prop_vax_to_se}{proportion of vaccines alloted to S & E compartments
(ie not R)}
}
\value{
state vector
}
\description{
Do vaccinations
}
