% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_inputs}
\alias{generate_inputs}
\title{Generate imports and vaccine inputs}
\usage{
generate_inputs(vaccine_doses = NULL, vaccine_days = NULL, import_days = NULL)
}
\arguments{
\item{vaccine_doses}{vector of vaccine counts}

\item{vaccine_days}{integer vector of time points, same length as
\code{vaccine_doses}}

\item{import_days}{integer vector of time points}
}
\value{
tibble, each row is an event (import or vaccination day)
}
\description{
Generate imports and vaccine inputs
}
