% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{rates}
\alias{rates}
\title{SEIR rates}
\usage{
rates(x, parameters, t)
}
\arguments{
\item{x}{state vector}

\item{parameters}{parameter grid row}

\item{t}{simulation time}
}
\value{
vector of rates
}
\description{
SEIR rates
}
