% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_params.R
\name{read_params}
\alias{read_params}
\title{Read simulation config file}
\usage{
read_params(path)
}
\arguments{
\item{path}{to yaml}
}
\value{
list of params
}
\description{
Config file must have:
\itemize{
\item \verb{$history}
\itemize{
\item \verb{$start_date}
\item \verb{$int_day}
\item \verb{$import_days}
\item \verb{$vaccine_days}
}
}
}
