% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_params.R
\name{replace_null}
\alias{replace_null}
\title{Replace NULL values by NA}
\usage{
replace_null(lst)
}
\arguments{
\item{lst}{a list}
}
\value{
vector with \code{NULL} values replaced by \code{NA}
}
\description{
Replace NULL values by NA
}
