% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_priors.R
\name{sample_priors}
\alias{sample_priors}
\title{Sample the prior distributions}
\usage{
sample_priors(n)
}
\arguments{
\item{n}{Number of samples to draw}
}
\value{
Named list of parameter values
}
\description{
Sample the prior distributions
}
\details{
r0 ~ pert(min = 10, mode = 20, max = 60, shape = 5)
int_eff ~ pert(min = 0, mode = .2, max = .8, shape = 5)
}
