% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seirMeasles-package.R
\docType{package}
\name{seirMeasles-package}
\alias{seirMeasles}
\alias{seirMeasles-package}
\title{seirMeasles: Stochastic Compartmental Model of Measles}
\description{
Modified version of Operation Allies Welcome model for 2024 Chicago measles outbreak.
}
\author{
\strong{Maintainer}: Scott Olesen \email{ulp7@cdc.gov}

Authors:
\itemize{
  \item Nina Masters \email{rhv2@cdc.gov}
  \item Inga Holmdahl \email{usn4@cdc.gov}
  \item Paige Miller \email{yub1@cdc.gov}
}

}
\keyword{internal}
