% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Run one simulation}
\usage{
simulate(params, manual_inputs, sim_length = 365, ssa_method = "exact", ...)
}
\arguments{
\item{params}{named vector of parameters}

\item{manual_inputs}{tibble as returned by \code{generate_inputs()}}

\item{sim_length}{duration of simulation in days (default: 365)}

\item{ssa_method}{passed to \code{method} argument of \code{\link[=ssa]{ssa()}}}

\item{...}{additional arguments passed to \code{\link[=ssa]{ssa()}}}
}
\value{
tibble
}
\description{
Run one simulation
}
