% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_config}
\alias{simulate_config}
\title{Run a grid of simulations, using a config setup}
\usage{
simulate_config(config, ...)
}
\arguments{
\item{config}{list}

\item{...}{Additional parameters passed to \code{\link[=simulate_grid]{simulate_grid()}}}
}
\description{
Run a grid of simulations, using a config setup
}
