% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_grid}
\alias{simulate_grid}
\title{Run simulations on a grid of parameters}
\usage{
simulate_grid(parameter_grid, manual_inputs, seed, ...)
}
\arguments{
\item{parameter_grid}{data.frame of values. Each row is interpreted as a
named list of scalar parameters}

\item{manual_inputs}{tibble of importations and vaccinations as produced
by \code{generate_inputs()}}

\item{seed}{random number seed}

\item{...}{Additional parameters passed to \code{\link[=simulate]{simulate()}}}
}
\value{
tibble of simulation results and parameters
}
\description{
Run simulations on a grid of parameters
}
