% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.R
\name{ssa}
\alias{ssa}
\title{Stochastic simulation algorithm (SSA)}
\usage{
ssa(
  state,
  transitions,
  rate_fun,
  end_time,
  start_time = 0,
  checkpoints = NULL,
  method = "exact",
  ...
)
}
\arguments{
\item{state}{named vector}

\item{transitions}{list of transitions passed to
\code{\link[adaptivetau:ssa.adaptivetau]{adaptivetau::ssa.adaptivetau()}}}

\item{rate_fun}{function of state and time, returning a vector of rates}

\item{end_time}{end of simulation}

\item{start_time}{default: 0}

\item{checkpoints}{vector of times; ensure rate_fun is re-evaluated at
these times}

\item{method}{one of \code{"exact"} (i.e., Gillespie; uses
\code{\link[adaptivetau:ssa.exact]{adaptivetau::ssa.exact()}}) or \code{"adaptive"} (uses
\code{\link[adaptivetau:ssa.adaptivetau]{adaptivetau::ssa.adaptivetau()}}). Default: \code{"exact"}}

\item{...}{additional parameters passed to \code{\link[adaptivetau:ssa.adaptivetau]{adaptivetau::ssa.adaptivetau()}}}
}
\value{
list of lists, one per transition. each list has elements \code{time} and
\code{state}
}
\description{
Wrap \code{\link[adaptivetau:ssa.exact]{adaptivetau::ssa.exact()}} or \code{\link[adaptivetau:ssa.adaptivetau]{adaptivetau::ssa.adaptivetau()}}, allowing
for checkpoints and specific start times
}
