% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\docType{data}
\name{transitions}
\alias{transitions}
\title{SEIR transitions}
\format{
An object of class \code{list} of length 26.
}
\usage{
transitions
}
\description{
See \code{\link[=ssa]{ssa()}} for details on transition matrix setup
}
\keyword{datasets}
