% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_priors.R
\name{sample_priors}
\alias{sample_priors}
\title{Sample the prior distributions}
\usage{
sample_priors(n)
}
\arguments{
\item{n}{Number of samples to draw}
}
\value{
Named list of parameter values
}
\description{
Sample the prior distributions
}
