% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Run one simulation}
\usage{
simulate(params, ...)
}
\arguments{
\item{params}{named list of parameters}

\item{...}{additional arguments passed to \code{\link[=ssa]{ssa()}}}
}
\value{
tibble
}
\description{
Run one simulation
}
